+++ lib/src/extensions.dart
@@ -43,3 +43,7 @@ extension NodeExt on Node {
     }
   }
 }
+
+extension ListExt<E> on List<E> {
+  E? get firstOrNull => isEmpty ? null : first;
+}