+++ example/main.dart
@@ -28,6 +28,8 @@ void main() {
   bs.find('p', class_: 'story'); // finds first element with html tag "p" and which has "class" attribute with value "story"
   bs.findAll('a', attrs: {'class': true}); // finds all elements with html tag "a" and which have defined "class" attribute with whatever value
   bs.find('*', id: 'link1'); // find by id
+  bs.find('*', regex: r'^b'); // find any elements that which tags starts with "b", for example: body, b, ...
+  bs.find('p', string: r'^Article #\d*'); // find "p" elements which texts starts with "Article #[number]"
   bs.find('a', attrs: {'href': 'http://example.com/elsie'}); // finds by "href" attribute
 
   // 3. perform any other actions for the navigated element